pyproject.toml
[project]
name = "disinformation_classification_pietervanbrakel"
version = "2.0.0"
description = "This project classifies news articles with the binary y predictions true and false."
readme = "README.md"
authors = [
  {name = "Pieter van Brakel", email = "pieter.van.brakel@hotmail.com"}
]
maintainers = [
  {name = "Pieter van Brakel", email = "pieter.van.brakel@hotmail.com"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "fastparquet>=2026.3.0",
  "typer",
  "pyyaml",
  "tqdm>=4.67.3",
  "scikit-learn==1.7.2",
  "nltk>=3.9.4",
  "matplotlib>=3.10.8",
  "seaborn>=0.13.2",
]
requires-python = ">=3.10,<3.12"

[project.optional-dependencies]
test = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "ipdb", # debugging
]

[project.urls]
bugs = "https://github.com/PieterVanBrakel/disinformation_classification_pietervanbrakel/issues"
changelog = "https://github.com/PieterVanBrakel/disinformation_classification_pietervanbrakel/blob/master/changelog.md"
homepage = "https://github.com/PieterVanBrakel/disinformation_classification_pietervanbrakel"

[project.scripts]
disinformation_classification_pietervanbrakel = "disinformation_classification_pietervanbrakel.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
